Added a couple of #include.
The newer compiler required that two #include be explicit.

Fixed the character struct, we can't have a constructor and the default
initializers are very similar to have a constructor. This is because the
struct is used with std::string and it is expected to be trivial and
constructors are not.